% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPDFilecheck.R
\name{calculate_age_from_dob}
\alias{calculate_age_from_dob}
\title{Function to calculate age from date of birth}
\usage{
calculate_age_from_dob(data, columnname, dateformat = "dmy", nrcode = NA)
}
\arguments{
\item{data}{a data frame}

\item{columnname}{name of column corresponding to date of birth}

\item{dateformat}{format of date e.g. dmy default is dmy}

\item{nrcode}{non response code corresponding to date of birth}
}
\value{
data if success error if failure
}
\description{
Function to calculate age from date of birth
}
\examples{
library(IPDFileCheck)
this.df <- data.frame(c("1987-05-28", "1987-06-18"), c(1, 2),
stringsAsFactors = FALSE)
colnames(this.df) <- c("dob", "num")
calculate_age_from_dob(this.df, "dob", "ymd")
}
