% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPDFilecheck.R
\name{calculate_age_from_year}
\alias{calculate_age_from_year}
\title{Function to calculate age from year of birth}
\usage{
calculate_age_from_year(data, columnname, nrcode = NA)
}
\arguments{
\item{data}{a data frame}

\item{columnname}{name of column corresponding to year of birth}

\item{nrcode}{non response code corresponding to date of birth}
}
\value{
data, if success error if failure
}
\description{
Function to calculate age from year of birth
}
\examples{
this.data.frame <- data.frame(c(1951, 1980), c("John", "Dora"))
colnames(this.data.frame) <- c("yob", "name")
calculate_age_from_year(this.data.frame, "yob", NA)
}
