% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPDdataFilecheck.R
\name{calculateAgeFromBirthYear}
\alias{calculateAgeFromBirthYear}
\title{Function to calculate age from year of birth}
\usage{
calculateAgeFromBirthYear(data, columnname, nrcode = NA)
}
\arguments{
\item{data}{a data frame}

\item{columnname}{name of column corresponding to year of birth}

\item{nrcode}{non response code corresponding to date of birth}
}
\value{
data, if success -1 if failure
}
\description{
Function to calculate age from year of birth
}
\examples{
calculateAgeFromBirthYear(data.frame("yob" = c(1951,1980), "Name" = c("John","Dora")),"yob",NA)
}
