% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPDdataFilecheck.R
\name{calculateAgeFromDob}
\alias{calculateAgeFromDob}
\title{Function to calculate age from date of birth}
\usage{
calculateAgeFromDob(data, columnname, dateformat = FALSE, nrcode = NA)
}
\arguments{
\item{data}{a data frame}

\item{columnname}{name of column corresponding to date of birth}

\item{dateformat}{format of date e.g. dmy default is FALSE}

\item{nrcode}{non response code corresponding to date of birth}
}
\value{
data if success -1 if failure
}
\description{
Function to calculate age from date of birth
}
\examples{
library(IPDFileCheck)
calculateAgeFromDob(data.frame("dob"=c("1987-05-28","1987-06-18","1987-07-09"),
"num"=c(1,2,3),stringsAsFactors = FALSE),"dob")
}
