% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPDFilecheck.R
\name{check_column_exists}
\alias{check_column_exists}
\title{Function to check the given column exists}
\usage{
check_column_exists(column_name, data)
}
\arguments{
\item{column_name}{a column name}

\item{data}{data frame}
}
\value{
0 if success error if failure
}
\description{
Function to check the given column exists
}
\examples{
check_column_exists("age", data.frame("Age" = c(21, 15),
"Name" = c("John", "Dora")))
}
