% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPDFilecheck.R
\name{check_load_packages}
\alias{check_load_packages}
\title{Function to check the package is installed, if not install}
\usage{
check_load_packages(pkg)
}
\arguments{
\item{pkg}{name of package(s)}
}
\value{
0, if packages cant be installed and loaded, else error
}
\description{
Function to check the package is installed, if not install
}
\examples{
check_load_packages("dplyr")
}
