% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPDdataFilecheck.R
\name{checkColumnExists}
\alias{checkColumnExists}
\title{Function to check the given column exists}
\usage{
checkColumnExists(column.name, data)
}
\arguments{
\item{column.name}{a column name}

\item{data}{data frame}
}
\value{
0 if success -1 if failure
}
\description{
Function to check the given column exists
}
\examples{
checkColumnExists("age",data.frame("Age" = c(21,15), "Name" = c("John","Dora")))
}
