% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPDdataFilecheck.R
\name{checkColumnNoForPatternInColumnname}
\alias{checkColumnNoForPatternInColumnname}
\title{Function to return the column number if a given pattern is contained in the column names of a data}
\usage{
checkColumnNoForPatternInColumnname(pattern, column.names)
}
\arguments{
\item{pattern}{a string that needs to be checked}

\item{column.names}{column names actually have}
}
\value{
TRUE , if success FALSE, if failure
}
\description{
Function to return the column number if a given pattern is contained in the column names of a data
}
\examples{
checkColumnNoForPatternInColumnname("age","female_age")
}
