% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPDdataFilecheck.R
\name{checkLoadPackages}
\alias{checkLoadPackages}
\title{Function to check the package is installed, if not install}
\usage{
checkLoadPackages(pkg)
}
\arguments{
\item{pkg}{name of package(s)}
}
\value{
0, if packages cant be installed and loaded, else -1
}
\description{
Function to check the package is installed, if not install
}
\examples{
checkLoadPackages("dplyr")
}
