% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPDdataFilecheck.R
\name{cohensD}
\alias{cohensD}
\title{Function to find the effect size Cohen's d}
\usage{
cohensD(x, y)
}
\arguments{
\item{x, }{a vector}

\item{y, }{another vector}
}
\value{
cohens d estimated with 95% CI or -1 if failure
}
\description{
Function to find the effect size Cohen's d
}
\examples{
cohensD(c(1,2,3,4),c(3,4,5,6))
}
