% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPDFilecheck.R
\name{convert_date_numeric_stdform}
\alias{convert_date_numeric_stdform}
\title{Helper function to keep date formats in year-month-date}
\usage{
convert_date_numeric_stdform(column, index, orderby = "dmy")
}
\arguments{
\item{column}{a data frame or a  vector}

\item{index}{those correspond to valid date in numeic form (omitting non
response code or no entry)}

\item{orderby}{give the order such as mdy, dmy etc where d refers to day,
m to month and y to year}
}
\value{
entry corrected entries as in standard date format
}
\description{
Helper function to keep date formats in year-month-date
}
\examples{
convert_date_numeric_stdform(c("01/01/2000", "02/02/2002"), c(1, 2), "dmy")
}
