% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPDFilecheck.R
\name{convert_to_number}
\alias{convert_to_number}
\title{Function that convert a number represented as character array}
\usage{
convert_to_number(character_array)
}
\arguments{
\item{character_array}{a character array of numbers}
}
\value{
converted_number in numeric form
}
\description{
Function that convert a number represented as character array
}
\examples{
convert_to_number(c("1","9","8"))
}
