% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPDdataFilecheck.R
\name{convertStdDateFormat}
\alias{convertStdDateFormat}
\title{Helper function to keep date formats in year/month/date}
\usage{
convertStdDateFormat(entry, index, monthfirst = NULL)
}
\arguments{
\item{entry}{a data frame or a  vector}

\item{index}{those correspond to valid date (omitting non response code or no entry)}

\item{monthfirst}{if month is given before date, NULL by default}
}
\value{
entry corrected entries
}
\description{
Helper function to keep date formats in year/month/date
}
\examples{
convertStdDateFormat(c("01/01/2000","02/02/2002"),c(1,2),NULL)
}
