% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPDFilecheck.R
\name{descriptive_stats_col}
\alias{descriptive_stats_col}
\title{Function to return descriptive statistics, sum, no of observations, mean,
mode. median, range, standard deviation and standard error}
\usage{
descriptive_stats_col(data, column_name, nrcode = NA)
}
\arguments{
\item{data}{data frame}

\item{column_name}{the column name}

\item{nrcode}{non response code corresponding to the column}
}
\value{
the descriptive statistics for success , error for failure
}
\description{
Function to return descriptive statistics, sum, no of observations, mean,
mode. median, range, standard deviation and standard error
}
\examples{
descriptive_stats_col(data.frame("age" = c(21, 15), "Name" = c("John", "Dora")), "age", NA)
}
