% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPDdataFilecheck.R
\name{descriptiveStatisticsDataColumn}
\alias{descriptiveStatisticsDataColumn}
\title{Function to return descriptive statistics, sum, no of observations, mean, mode. median, range, standard deviation and standard error}
\usage{
descriptiveStatisticsDataColumn(data, column.name, nrcode = NA)
}
\arguments{
\item{data}{data frame}

\item{column.name}{the column name}

\item{nrcode}{non response code corresponding to the column}
}
\value{
the descriptive statistics for success , -1 for failure
}
\description{
Function to return descriptive statistics, sum, no of observations, mean, mode. median, range, standard deviation and standard error
}
\examples{
descriptiveStatisticsDataColumn(data.frame("Age" = c(21,15), "Name" = c("John","Dora")),"age",NA)
}
