% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPDFilecheck.R
\name{get_colno_pattern_colname}
\alias{get_colno_pattern_colname}
\title{Function to return the column number if a given pattern is contained
in the column names of a data}
\usage{
get_colno_pattern_colname(pattern, column_names)
}
\arguments{
\item{pattern}{a string that needs to be checked}

\item{column_names}{column names actually have}
}
\value{
column number, if success error, if failure
}
\description{
Function to return the column number if a given pattern is contained
in the column names of a data
}
\examples{
get_colno_pattern_colname("age", "female_age")
}
