% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPDFilecheck.R
\name{get_contents_cols}
\alias{get_contents_cols}
\title{Function to return the unique contents of the column given the column name}
\usage{
get_contents_cols(data, colname)
}
\arguments{
\item{data}{a data frame}

\item{colname}{name of column corresponding to year of birth}
}
\value{
the contents of the column, if success error if failure
}
\description{
Function to return the unique contents of the column given the column name
}
\examples{
get_contents_cols(data.frame("yob" =  c(1951,1980),
"Name" =  c("John","Dora")),"yob")
}
