% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPDFilecheck.R
\name{get_effect_size}
\alias{get_effect_size}
\title{Function to get the effect size}
\usage{
get_effect_size(data, variable, by, ...)
}
\arguments{
\item{data}{a data frame}

\item{variable}{variables to be selected for summary}

\item{by}{A column name (quoted or unquoted) in data.}

\item{...}{extra parameters required}
}
\value{
returns the effect sizes
}
\description{
Function to get the effect size
}
