% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPDFilecheck.R
\name{get_mode_from_vector}
\alias{get_mode_from_vector}
\title{############################################################################
Function to return mode}
\usage{
get_mode_from_vector(v)
}
\arguments{
\item{v}{a vector}
}
\value{
mode
}
\description{
############################################################################
Function to return mode
}
\examples{
get_mode_from_vector(c(1, 1, 2, 3))
}
