% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPDFilecheck.R
\name{get_sem}
\alias{get_sem}
\title{###############################################################################
Function to estimate standard error of the mean}
\usage{
get_sem(x)
}
\arguments{
\item{x, }{a vector}
}
\value{
SE the standard error of the mean
}
\description{
###############################################################################
Function to estimate standard error of the mean
}
\examples{
get_sem(c(1, 2, 3, 4))
}
