% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPDFilecheck.R
\name{get_summary_gtsummary}
\alias{get_summary_gtsummary}
\title{Function to return the summary table using gtsummary package}
\usage{
get_summary_gtsummary(the_data, selectvar, byvar = NULL, label = NULL)
}
\arguments{
\item{the_data}{a data frame}

\item{selectvar}{variables to be selected for summary}

\item{byvar}{A column name (quoted or unquoted) in data.
Summary statistics will be calculated separately for each level
of the by variable. If NULL, summary statistics are calculated
using all observations.}

\item{label}{List of formulas specifying variables labels,}
}
\value{
the summary using gtsummarys  tbl_summary option
}
\description{
Function to return the summary table using gtsummary package
}
\examples{
trial <- gtsummary::trial
table1 <- get_summary_gtsummary(trial, c("trt", "age", "grade"),
byvar = "trt")
}
