% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPDFilecheck.R
\name{get_value_from_codes}
\alias{get_value_from_codes}
\title{Function to get the actual value of column content if its coded}
\usage{
get_value_from_codes(data, column, nrcode = NA, list_codes_values)
}
\arguments{
\item{data}{a data frame}

\item{column}{column name for value}

\item{nrcode}{non response code corresponding to gender column}

\item{list_codes_values}{list of codes to understand the codes and value}
}
\value{
0, if success error if failure
}
\description{
Function to get the actual value of column content if its coded
}
\examples{
data = data.frame("sex" = c(1, 2, 2, 1, 1),
 "Name" = c("John", "Dora","Dora", "John","John"))
list_codes_values = list(c("F", "M"),c(1,2))
ans <- get_value_from_codes(data, column = "sex", nrcode = NA,
list_codes_values)
}
