% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPDdataFilecheck.R
\name{getColumnNoForNames}
\alias{getColumnNoForNames}
\title{Function to return the column number for column name}
\usage{
getColumnNoForNames(data, column.name)
}
\arguments{
\item{data}{a data frame}

\item{column.name}{column names of the data frame}
}
\value{
column number, if success -1, if failure
}
\description{
Function to return the column number for column name
}
\examples{
getColumnNoForNames(data.frame("Age"= c(21,15), "Name"= c("John","Dora")),"Name")
}
