% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPDdataFilecheck.R
\name{getColumnNoForPatternInColumnname}
\alias{getColumnNoForPatternInColumnname}
\title{Function to return the column number if a given attern is contained in the column names of a data}
\usage{
getColumnNoForPatternInColumnname(pattern, column.names)
}
\arguments{
\item{pattern}{a string that needs to be checked}

\item{column.names}{column names actually have}
}
\value{
column number, if success -1, if failure
}
\description{
Function to return the column number if a given attern is contained in the column names of a data
}
\examples{
getColumnNoForPatternInColumnname("age","female_age")
}
