% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPDdataFilecheck.R
\name{getConentdInCols}
\alias{getConentdInCols}
\title{Function to return the unique contents of the column given the column name}
\usage{
getConentdInCols(data, colname)
}
\arguments{
\item{data}{a data frame}

\item{colname}{name of column corresponding to year of birth}
}
\value{
the contents of the column, if success -1 if failure
}
\description{
Function to return the unique contents of the column given the column name
}
\examples{
getConentdInCols(data.frame("yob" = c(1951,1980), "Name" = c("John","Dora")),"yob")
}
