% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPDdataFilecheck.R
\name{getModeForVector}
\alias{getModeForVector}
\title{Function to return mode}
\usage{
getModeForVector(v)
}
\arguments{
\item{v}{a vector}
}
\value{
mode
}
\description{
Function to return mode
}
\examples{
getModeForVector(c(1,1,2,3))
}
