% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPDdataFilecheck.R
\name{getSEM}
\alias{getSEM}
\title{Function to estimate standard error of the mean}
\usage{
getSEM(x)
}
\arguments{
\item{x, }{a vector}
}
\value{
SE the standard error of the mean
}
\description{
Function to estimate standard error of the mean
}
\examples{
getSEM(c(1,2,3,4))
}
