% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPDFilecheck.R
\name{present_mean_sd_rmna_text}
\alias{present_mean_sd_rmna_text}
\title{############################################################################
Function to present the mean and sd of a data set in the form Mean (SD)}
\usage{
present_mean_sd_rmna_text(data, column_name, nrcode = NA)
}
\arguments{
\item{data}{data frame}

\item{column_name}{the column name}

\item{nrcode}{non response code corresponding to the column}
}
\value{
the mean(sd), error for failure
}
\description{
############################################################################
Function to present the mean and sd of a data set in the form Mean (SD)
}
\examples{
present_mean_sd_rmna_text(data.frame(
  "age" = c(21, 15),
  "Name" = c("John", "Dora")
), "age", NA)
}
