% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPDdataFilecheck.R
\name{presentMeanSdRemoveNAText}
\alias{presentMeanSdRemoveNAText}
\title{Function to present the mean and sd of a data set in the form Mean (SD)}
\usage{
presentMeanSdRemoveNAText(data, column.name, nrcode = NA)
}
\arguments{
\item{data}{data frame}

\item{column.name}{the column name}

\item{nrcode}{non response code corresponding to the column}
}
\value{
the mean(sd), -1 for failure
}
\description{
Function to present the mean and sd of a data set in the form Mean (SD)
}
\examples{
presentMeanSdRemoveNAText(data.frame("Age" = c(21,15), "Name" = c("John","Dora")),"age",NA)
}
