% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPDFilecheck.R
\name{represent_categorical_data}
\alias{represent_categorical_data}
\title{Function to find the number and percentages of categories}
\usage{
represent_categorical_data(data, variable, nrcode = NA)
}
\arguments{
\item{data, }{a data frame}

\item{variable}{the column name}

\item{nrcode}{non response code}
}
\value{
number and percentages or error if failure
}
\description{
Function to find the number and percentages of categories
}
\examples{
this.df <- data.frame(c(11, 78), c("m", "f"), stringsAsFactors = FALSE)
colnames(this.df) <- c("mark", "gender")
represent_categorical_data(this.df, "gender", NA)
}
