% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPDFilecheck.R
\name{represent_categorical_data_forsubgroups}
\alias{represent_categorical_data_forsubgroups}
\title{Function to find the number and percentages of categories}
\usage{
represent_categorical_data_forsubgroups(
  data,
  variable1,
  variable2,
  nrcode = NA
)
}
\arguments{
\item{data, }{a data frame}

\item{variable1}{the column name of the variable to be grouped based on}

\item{variable2}{the column name of the variable to represented}

\item{nrcode}{non response code for the variable2}
}
\value{
the subgroup
}
\description{
Function to find the number and percentages of categories
}
\examples{
this.df <- data.frame(c(11, 78,22), c("m", "f", "f"), c(1,2,2),
stringsAsFactors = FALSE)
colnames(this.df) <- c("mark", "gender", "group")
represent_categorical_data_forsubgroups(this.df, "group", "gender", NA)
}
