% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPDFilecheck.R
\name{represent_categorical_textdata}
\alias{represent_categorical_textdata}
\title{Function to represent categorical data in the form - numbers (percentage)}
\usage{
represent_categorical_textdata(data, variable, nrcode)
}
\arguments{
\item{data}{data frame}

\item{variable}{column name}

\item{nrcode}{non response code}
}
\value{
the numbers (percentage) , error for failure
}
\description{
Function to represent categorical data in the form - numbers (percentage)
}
\examples{
df <- data.frame(c(11, 78), c("m", "f"), stringsAsFactors = FALSE)
colnames(df) <- c("mark", "gender")
represent_categorical_textdata(df, "gender", NA)
}
