% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPDdataFilecheck.R
\name{representCategoricalData}
\alias{representCategoricalData}
\title{Function to find the number and percentages of categories}
\usage{
representCategoricalData(data, variable, nrcode = NA)
}
\arguments{
\item{data, }{a data frame}

\item{variable}{the column name}

\item{nrcode}{non response code}
}
\value{
number and percentages or -1 if failure
}
\description{
Function to find the number and percentages of categories
}
\examples{
representCategoricalData(data.frame(mark=c(0,11,78,160),gender=c("f","m","f","m")),"gender",NA)
}
