% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPDdataFilecheck.R
\name{representCategoricalDataText}
\alias{representCategoricalDataText}
\title{Function to represent categorical data in the form - numbers (percentage)}
\usage{
representCategoricalDataText(data, variable, nrcode)
}
\arguments{
\item{data}{data frame}

\item{variable}{column name}

\item{nrcode}{non response code}
}
\value{
the numbers (percentage) , -1 for failure
}
\description{
Function to represent categorical data in the form - numbers (percentage)
}
\examples{
representCategoricalDataText(data.frame(mark=c(0,11,78,160),gender=c("f","m","f","m")),"gender",NA)
}
