% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPDFilecheck.R
\name{return_longitudinal_summary}
\alias{return_longitudinal_summary}
\title{Function to get the longitudinal summary mean and sd}
\usage{
return_longitudinal_summary(thedata, columnnames, nrcode = NA)
}
\arguments{
\item{thedata}{a data frame}

\item{columnnames}{column names of the data that are some observations
at some time points}

\item{nrcode}{the non response code in the data}
}
\value{
returns the effect sizes
}
\description{
Function to get the longitudinal summary mean and sd
}
\examples{
test_data <- as.data.frame(cbind(c(1,2,3,4,5), c(20,40,60,80,100),
c("F", "F", "M", "M", "F")))
colnames(test_data) <- c("no", "marks", "gender")
test_data$marks <- as.numeric(test_data$marks)
results <- return_longitudinal_summary(test_data, "marks", NA)
}
