% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPDFilecheck.R
\name{return_subgroup_withNA}
\alias{return_subgroup_withNA}
\title{Function to return a subgroup when certain variable equals the given value
while omitting those with NA}
\usage{
return_subgroup_withNA(data, variable, value)
}
\arguments{
\item{data}{data frame}

\item{variable}{that corresponds to a column}

\item{value}{a value that can be taken by the variable}
}
\value{
subgroup a data frame if success error if failure
}
\description{
Function to return a subgroup when certain variable equals the given value
while omitting those with NA
}
\examples{
return_subgroup_withNA(data.frame(
  "age" = c(21, 15),
  "Name" = c("John", "Dora")
), "age", 10)
}
