% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPDdataFilecheck.R
\name{returnSubgroupOmitNA}
\alias{returnSubgroupOmitNA}
\title{Function to return a subgroup when certain variable equals the given value while omitting those with NA}
\usage{
returnSubgroupOmitNA(data, variable, value)
}
\arguments{
\item{data}{data frame}

\item{variable}{that corresponds to a column}

\item{value}{a value that can be taken by the variable}
}
\value{
subgroup a data frame if success -1 if failure
}
\description{
Function to return a subgroup when certain variable equals the given value while omitting those with NA
}
\examples{
returnSubgroupOmitNA(data.frame("age" = c(21,15), "Name" = c("John","Dora")),"age",10)
}
