% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPDFilecheck.R
\name{test_age}
\alias{test_age}
\title{Function to check the format of 'age' in data}
\usage{
test_age(data, agecolumn = "age", nrcode = NA)
}
\arguments{
\item{data}{a data frame}

\item{agecolumn}{column name that corresponds to age or date pf birth}

\item{nrcode}{non response code corresponding to age column}
}
\value{
0, if success error if failure
}
\description{
Function to check the format of 'age' in data
}
\examples{
df <- data.frame("Age" = c(21, 15), "Name" = c("John", "Dora"))
test_age(df, "age", 999)
}
