% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPDFilecheck.R
\name{test_column_contents}
\alias{test_column_contents}
\title{Function to check the format of column contents}
\usage{
test_column_contents(data, column, code, nrcode = NA)
}
\arguments{
\item{data}{a data frame}

\item{column}{column name for gender}

\item{code}{how column values are  coded}

\item{nrcode}{non response code corresponding to gender column}
}
\value{
0, if success error if failure
}
\description{
Function to check the format of column contents
}
\examples{
test_column_contents(data.frame("sex" =  c("m","f"),
"Name"=  c("John","Dora")),"sex",c("m","f"),999)
}
