% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPDFilecheck.R
\name{test_columnnames}
\alias{test_columnnames}
\title{Function to test column names of a data being different from what specified}
\usage{
test_columnnames(column_names, data)
}
\arguments{
\item{column_names}{column names of the data frame}

\item{data}{a data frame}
}
\value{
0, if success error, if failure
}
\description{
Function to test column names of a data being different from what specified
}
\examples{
test_columnnames(c("name","age"), data.frame("Age" =  c(21,15),
"Name" =  c("John","Dora")))
}
