% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPDFilecheck.R
\name{test_data_numeric}
\alias{test_data_numeric}
\title{Function to check the format of a numeric column}
\usage{
test_data_numeric(column_name, data, nrcode = NA, minval, maxval)
}
\arguments{
\item{column_name}{the column name}

\item{data}{data frame}

\item{nrcode}{non response code corresponding to the column}

\item{minval}{minimum value allowed}

\item{maxval}{maximum value allowed}
}
\value{
0, if success error, if failure
}
\description{
Function to check the format of a numeric column
}
\examples{
test_data_numeric("age", data.frame(
  "Age" =  c(21, 15),
  "Name"  = c("John", "Dora")
), -99, 0, 100)
}
