% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPDFilecheck.R
\name{test_data_numeric_norange}
\alias{test_data_numeric_norange}
\title{Function to check the format of a numeric column when the values
are not bounded}
\usage{
test_data_numeric_norange(column_name, data, nrcode = NA)
}
\arguments{
\item{column_name}{the column name}

\item{data}{data frame}

\item{nrcode}{non response code corresponding to the column}
}
\value{
0, if success error, if failure
}
\description{
Function to check the format of a numeric column when the values
are not bounded
}
\examples{
test_data_numeric_norange("marks", data.frame(
  "marks"  =  c(210, 99),
  "Name" =  c("John", "Dora")
), -99)
}
