% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPDFilecheck.R
\name{test_data_string}
\alias{test_data_string}
\title{Function to check the format of a string column}
\usage{
test_data_string(data, column_name, nrcode = NA)
}
\arguments{
\item{data}{data frame}

\item{column_name}{the column name}

\item{nrcode}{non response code corresponding to the column}
}
\value{
0, if success error, if failure
}
\description{
Function to check the format of a string column
}
\examples{
test_data_string(data.frame("Age" = c(21,15), "Name" = c("John","Dora")),"name",-999)
}
