% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPDFilecheck.R
\name{test_data_string_restriction}
\alias{test_data_string_restriction}
\title{Function to check the format of a string column when the string values are given}
\usage{
test_data_string_restriction(data, column_name, nrcode = NA,
  allowed_strings)
}
\arguments{
\item{data}{data frame}

\item{column_name}{the column name}

\item{nrcode}{non response code corresponding to the column}

\item{allowed_strings}{allowed strings or characters to represent meaningful entry}
}
\value{
0, if success error, if failure
}
\description{
Function to check the format of a string column when the string values are given
}
\examples{
test_data_string_restriction(data.frame("Age" = c(21,15), "sex" =  c("m","f")),
"sex",-999,c("f","m"))
}
