% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPDFilecheck.R
\name{test_file_exist_read}
\alias{test_file_exist_read}
\title{Function to throw error on invalid directory or file and if not readable}
\usage{
test_file_exist_read(filename)
}
\arguments{
\item{filename}{name of a file or dir}
}
\value{
0, if success error, if failure
}
\description{
Function to throw error on invalid directory or file and if not readable
}
\examples{
test_file_exist_read(system.file("extdata", "blank.txt",
package = "IPDFileCheck"))
}
