% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPDFilecheck.R
\name{test_gender}
\alias{test_gender}
\title{Function to check the format of 'gender' column in data}
\usage{
test_gender(data, gendercode, gendercolumn = "gender", nrcode = NA)
}
\arguments{
\item{data}{a data frame}

\item{gendercode}{how gender is coded}

\item{gendercolumn}{column name for gender}

\item{nrcode}{non response code corresponding to gender column}
}
\value{
0, if success error if failure
}
\description{
Function to check the format of 'gender' column in data
}
\examples{
test_gender(data.frame("sex" =  c("m","f"), "Name" =  c("John","Dora")),c("f", "m"),"sex",999)
}
