% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPDdataFilecheck.R
\name{testAge}
\alias{testAge}
\title{Function to check the format of 'age' in data}
\usage{
testAge(data, agecolumn = "age", nrcode = NA)
}
\arguments{
\item{data}{a data frame}

\item{agecolumn}{column name that corresponds to age or date pf birth}

\item{nrcode}{non response code corresponding to age column}
}
\value{
0, if success -1 if failure
}
\description{
Function to check the format of 'age' in data
}
\examples{
testAge(data.frame("Age" = c(21,15), "Name" = c("John","Dora")),"age",999)
}
