% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPDdataFilecheck.R
\name{testColumnContents}
\alias{testColumnContents}
\title{Function to check the format of column contents}
\usage{
testColumnContents(data, column, code, nrcode = NA)
}
\arguments{
\item{data}{a data frame}

\item{column}{column name for gender}

\item{code}{how column values are  coded}

\item{nrcode}{non response code corresponding to gender column}
}
\value{
0, if success -1 if failure
}
\description{
Function to check the format of column contents
}
\examples{
testColumnContents(data.frame("sex" = c("m","f"), "Name" = c("John","Dora")),"sex",c("m","f"),999)
}
