% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPDdataFilecheck.R
\name{testDataColumnNames}
\alias{testDataColumnNames}
\title{Function to test column names of a data being different from what specified}
\usage{
testDataColumnNames(column.names, data)
}
\arguments{
\item{column.names}{column names of the data frame}

\item{data}{a data frame}
}
\value{
0, if success -1, if failure
}
\description{
Function to test column names of a data being different from what specified
}
\examples{
testDataColumnNames(c("name","age"),data.frame("Age" = c(21,15), "Name" = c("John","Dora")))
}
