% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPDdataFilecheck.R
\name{testDataNumeric}
\alias{testDataNumeric}
\title{Function to check the format of a numeric column}
\usage{
testDataNumeric(column.name, data, nrcode = NA, minval, maxval)
}
\arguments{
\item{column.name}{the column name}

\item{data}{data frame}

\item{nrcode}{non response code corresponding to the column}

\item{minval}{minimum value allowed}

\item{maxval}{maximum value allowed}
}
\value{
0, if success -1, if failure
}
\description{
Function to check the format of a numeric column
}
\examples{
testDataNumeric("age",data.frame("Age" = c(21,15), "Name" = c("John","Dora")),-99,0,100)
}
