% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPDdataFilecheck.R
\name{testDataNumericNorange}
\alias{testDataNumericNorange}
\title{Function to check the format of a numeric column when the values are not bounded}
\usage{
testDataNumericNorange(column.name, data, nrcode = NA)
}
\arguments{
\item{column.name}{the column name}

\item{data}{data frame}

\item{nrcode}{non response code corresponding to the column}
}
\value{
0, if success -1, if failure
}
\description{
Function to check the format of a numeric column when the values are not bounded
}
\examples{
testDataNumericNorange("marks",data.frame("marks" = c(210,99), "Name" = c("John","Dora")),-99)
}
