% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPDdataFilecheck.R
\name{testDataString}
\alias{testDataString}
\title{Function to check the format of a string column}
\usage{
testDataString(data, column.name, nrcode = NA)
}
\arguments{
\item{data}{data frame}

\item{column.name}{the column name}

\item{nrcode}{non response code corresponding to the column}
}
\value{
0, if success -1, if failure
}
\description{
Function to check the format of a string column
}
\examples{
testDataString(data.frame("Age" = c(21,15), "Name" = c("John","Dora")),"name",-999)
}
