% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPDdataFilecheck.R
\name{testDataStringRestriction}
\alias{testDataStringRestriction}
\title{Function to check the format of a string column when the string values are given}
\usage{
testDataStringRestriction(data, column.name, nrcode = NA,
  allowed.strings)
}
\arguments{
\item{data}{data frame}

\item{column.name}{the column name}

\item{nrcode}{non response code corresponding to the column}

\item{allowed.strings}{allowed strings or characters to represent meaningful entry}
}
\value{
0, if success -1, if failure
}
\description{
Function to check the format of a string column when the string values are given
}
\examples{
testDataStringRestriction(data.frame("Age" = c(21,15), "sex" = c("m","f")),"sex",-999,c("f","m"))
}
