% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPDdataFilecheck.R
\name{testFileExistenceReadability}
\alias{testFileExistenceReadability}
\title{Function to throw error on invalid directory or file or if the file is not readable}
\usage{
testFileExistenceReadability(filename)
}
\arguments{
\item{filename}{name of a file or dir}
}
\value{
0, if success -1, if failure
}
\description{
Function to throw error on invalid directory or file or if the file is not readable
}
\examples{
testFileExistenceReadability(system.file("extdata", "blank.txt", package = "IPDFileCheck"))
}
