% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPDdataFilecheck.R
\name{testGender}
\alias{testGender}
\title{Function to check the format of 'gender' column in data}
\usage{
testGender(data, gendercode, gendercolumn = "gender", nrcode = NA)
}
\arguments{
\item{data}{a data frame}

\item{gendercode}{how gender is coded}

\item{gendercolumn}{column name for gender}

\item{nrcode}{non response code corresponding to gender column}
}
\value{
0, if success -1 if failure
}
\description{
Function to check the format of 'gender' column in data
}
\examples{
testGender(data.frame("sex"= c("m","f"), "Name"= c("John","Dora")),c("f", "m"),"sex",999)
}
