% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPDFilecheck.R
\name{return_longitudinal_summary}
\alias{return_longitudinal_summary}
\title{Function to get the longitudinal summary mean and sd}
\usage{
return_longitudinal_summary(thedata, columnnames, nrcode = NA)
}
\arguments{
\item{thedata}{a data frame}

\item{columnnames}{column names of the data that are some observations
at some time points}

\item{nrcode}{the non response code in the data}
}
\value{
returns the effect sizes
}
\description{
Function to get the longitudinal summary mean and sd
}
